+++ Sources/SingBridgeCore/NetworkDefaults.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public enum SingBridgeNetworkDefaults {
+  public static let serviceName = "SingBridge"
+  public static let serviceType = "_singbridge._tcp"
+  public static let port: UInt16 = 49_555
+}